requirements.txt
pandas
requests
pytz
python-telegram-bot
openpyxl
opencage
